// ansteype-front/src/app/docs/s022/models/s022Dto.ts
export interface Child {
  is_del:boolean;
  // 統合
  integration_selected:boolean;
  // 月払
  monthly_payment_selected:boolean;
  id: number;
  matter_master_exe_id: number;
  no: string;
  sb_store: string;
  agency_id: number;
  agency_name: string;
  participating_store: string;
  weekday_weekend :  string;    // 一時保存フラグ
  held_date: string;
  event_location: string;
  order: string;
  number_people: string;
  item_name: string;
  type_flg: string;
  // 催事場代
  event_venue_fee: string;
  // 交通費
  transportation_fee: string;
  isEditable: boolean;
  days: string;

  // ページ分け
  page_division_1_selected:boolean;
  page_division_2_selected:boolean;
  // 非課税
  tax_free_selected:boolean;
  // googleEXCELファイル名
  google_excel_name: string;
  // googleEXCELファイルシート名
  google_excel_sheet_name: string;
  // 一時保存フラグ
  temporary_save_flg: string;
  conditions_year: string;
  conditions_month: string;
  conditions_week: string;
  conditions_agency: string;
}
export  interface Product {

    id: number;
    matter_master_exe_id: number;
    agency_id: number;
    agency_name: string;
    participating_store: string;
    isEditable: false;
    // 開催日
    held_date: string;
    // イベント場所
    event_location: string;

    // googleEXCELファイル名
    google_excel_name: string;
    // googleEXCELファイルシート名
    google_excel_sheet_name: string;
    conditions_year: string;
    conditions_month: string;
    conditions_week: string;
    conditions_agency: string;
    children: Child[];  // 使用 Child 类型数组来表示父数据的子项

}
// ...
  export interface  S022SeachDto {
    year:  string;
    month:  string;
    week:  string;
    // 代理店
    agency: string;
  }
// ...
 export interface  reDto{
    s003SearchDto:S003SeachDto;
    products: Product[];
  }
  export interface  reDtoSave{
    s003SearchDto:S003SeachDto;
    products: Product[];
    exeName: string;
  }

  export interface  reDtodDelSave{
    s003SearchDto:S003SeachDto;
    products: Product[];
    temporarySaveInfoDto:TemporarySaveInfoDto;
    exeName: string;
  }